+++ Code/CryEngine/RenderDll/Common/RendElements/IrisShafts.cpp
@@ -228,7 +228,7 @@ bool IrisShafts::PreparePrimitives(const SPreparePrimitivesContext& context)
 			m_fAngleRange = 0.65f * occ * occ + 0.35f;
 			m_fPrimaryDir = occQuery->GetDirResult() / (2 * PI);
 			m_fConcentrationBoost = 2.0f - occ;
-			m_fBrightnessBoost = 1 + 2 * pow(occ - 1, 6);
+			m_fBrightnessBoost = static_cast<float>(1 + 2 * pow(occ - 1, 6));
 
 			m_meshDirty = true;
 			m_fPrevOcc = occ;
@@ -236,7 +236,7 @@ bool IrisShafts::PreparePrimitives(const SPreparePrimitivesContext& context)
 		else if (occ < 0.05f)
 		{
 			m_fAngleRange = 0.33f;
-			m_fBrightnessBoost = 1 + 2 * pow(interpOcc - 1, 6);
+			m_fBrightnessBoost = static_cast<float>(1 + 2 * pow(interpOcc - 1, 6));
 			m_fPrevOcc = 0;
 			m_meshDirty = true;
 		}
